# docs/openapi-web.yaml
openapi: 3.1.0
...
info:
  title: TokenMP Web 页面通信
  version: "0"
  description: |
    页面通信与数据面的分界：

    - 本文件（`/api/v1/*`）：浏览器页面调用的会话、账户、业务数据端点；
    - `docs/openapi.yaml`（`/v1/*`、`/v1beta/*`、`/healthz`）：模型转发与账户自助查询，
      由模型客户端与集成调用，页面不直接调用。

    信封语义：

    - `code` 业务结果码，程序按它分支：`200` 成功，其余为失败；
      HTTP 状态码表达传输层结果，两者在成功时一致，失败时以 `code` 为准；
      `message` 面向用户的文案，程序不得按它分支。
    - `data` 业务数据，结构由各端点声明；无数据时省略或 `null`。
    - `page` / `size` / `total` 分页信息，仅列表端点填充，其余端点为 `null`。
      `page` 从 1 起，`size` 为每页条数，`total` 为总条数。

    业务码取值：

    | code | 含义 |
    |---|---|
    | 200 | 成功 |
    | 400 | 参数缺失或非法 |
    | 401 | 未登录、会话失效或凭据错误 |
    | 402 | 账户无可用额度，按账户页的存量与引导区块充值或购买 |
    | 403 | 已登录但无权访问 |
    | 404 | 资源不存在或登录方式未启用 |
    | 409 | 状态冲突（邮箱已注册等） |
    | 410 | 一次性加密公钥失效，重新获取后重试 |
    | 429 | 触发频率限制 |
    | 500 | 服务端错误 |

    密码不传明文，流程按客户端指纹串起：

    1. 客户端持有本地生成的 `fingerprint`（在本地持久化，随每次请求带回）；
    2. 调 `GET /api/v1/auth/challenge?fingerprint=…` 换取一次性公钥，
       服务端以该指纹为索引保存私钥，同一指纹再次获取会覆盖旧值；
    3. 用 RSA-OAEP（SHA-256）加密密码，提交密文与同一 `fingerprint`；
    4. 私钥取出即失效，密钥过期或已用返回 `code=410`，客户端重新执行 2-3 后重试一次。

    登录方式：账号密码（signin / signup）与第三方 OAuth（Google、GitHub 等，
    经 `providers` 列出、`oauth` 跳转、`exchange` 换取会话）。注销账号（erase）
    与重置密码（reset）各自需要对应用途的 OTP。

    路径分区：

    - `/api/v1/auth/*`：认证面，注册、登录、会话、密码与第三方登录。
    - `/api/v1/user/*`：用户级业务数据。作用域由会话推导，不接受账户 id 参数；
      一个登录主体拥有一个账户（`account.owner_user_id`），密钥、用量与请求记录
      都挂在该账户上。列表端点用偏移分页，分页信息在信封的 `page` / `size` / `total`。
...
servers:
...
security:
  - webSession: []
paths:
...
  /api/v1/admin/merchants:
    get:
      tags: [admin]
      summary: 商家清单
      description: |
        列出全部商家，与 `tokenmp admin merchant list --json` 同一份行数据。
        渠道与凭据的写入都要给出归属商家，页面据此取候选；平台自营与入驻商家
        都是这里的行，`kind` 只是数据。请求方须持有管理面能力，越权返回 `code=403`。
      operationId: listAdminMerchants
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: 商家清单。
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PageOfAdminMerchant"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [admin]
      summary: 新建商家
      description: |
        新建一个商家，与 `tokenmp admin merchant create` 调用同一份业务实现。新建一律
        `active`，停用是独立动作。平台自营与入驻商家走同一条路径，`kind` 只是数据。
        请求方须持有管理面能力，越权返回 `code=403`。
      operationId: createAdminMerchant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdminMerchantRequest"
      responses:
        "200":
          description: 新建成功，`data.id` 是新行主键。
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AdminResourceID"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/admin/channels:
    get:
      tags: [admin]
      summary: 渠道清单
      description: |
        列出全部上游渠道，与 `tokenmp admin channel list --json` 同一份行数据。
        渠道的 `config` 不回显：探针请求头里可能有鉴权 token。
        请求方须持有管理面能力，越权返回 `code=403`。
      operationId: listAdminChannels
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: 渠道清单。
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PageOfAdminChannel"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [admin]
      summary: 新建渠道
      description: |
        新建一条上游渠道，与 `tokenmp admin channel create` 调用同一份业务实现。
        新建一律启用，启用位由后续动作改动。`priority` / `weight` 缺省取 100：0 会让渠道
        永远排在最后，因此不作为「未配置」的取值。`cred_group` 只是分组名，凭据要另行写入。

        `config` 是渠道级扩展配置原文（静态请求头与探针声明），必须是 JSON 对象；
        非对象、`headers` 不是字符串到字符串的对象、或占用了网关自身请求头的头名，
        都在写入时返回 `code=400`（读取侧对这些情况是宽容的，写入口是唯一能显式拦下的地方）。

        请求方须持有管理面能力，越权返回 `code=403`。
      operationId: createAdminChannel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdminChannelRequest"
      responses:
        "200":
          description: 新建成功，`data.id` 是新行主键。
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AdminResourceID"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/admin/credentials:
    get:
      tags: [admin]
      summary: 上游凭据清单
      description: |
        列出全部上游凭据，与 `tokenmp admin credential list --json` 同一份行数据。
        只给脱敏前缀与形态，明文与 secret JSON 任何端点都不返回。
        请求方须持有管理面能力，越权返回 `code=403`。
      operationId: listAdminCredentials
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: 凭据清单。
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PageOfAdminCredential"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [admin]
      summary: 写入上游凭据
      description: |
        写入一行上游凭据，与 `tokenmp admin credential add` 调用同一份业务实现，
        secret 以 `{"api_key": …}` 承载。请求体里的 `api_key` 是明文，只在写入这一条
        路径出现；写入响应不回显明文，读取永远只有脱敏前缀。

        订阅型上游的 OAuth 登录需要设备码轮询等交互过程，不是单次请求能完成的动作，
        不在本端点：它仍走 `tokenmp admin credential oauth-login`。

        请求方须持有管理面能力，越权返回 `code=403`。
      operationId: createAdminCredential
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdminCredentialRequest"
      responses:
        "200":
          description: 写入成功，`data.id` 是新行主键。
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AdminResourceID"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/admin/modelmaps:
    get:
      tags: [admin]
      summary: 渠道模型映射清单
      description: |
        列出全部渠道模型映射，与 `tokenmp admin model-map list --json` 同一份行数据。
        路径段用单词（`modelmaps`），与认证面路径的命名约定一致。
        请求方须持有管理面能力，越权返回 `code=403`。
      operationId: listAdminModelMaps
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: 模型映射清单。
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PageOfAdminModelMap"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      tags: [admin]
      summary: 写入渠道模型映射
      description: |
        写入或覆盖一条渠道模型映射，与 `tokenmp admin model-map set` 调用同一份业务实现。
        写入按 `(channel_id, model)` 幂等：同一组合重复写入只覆盖取值，不新增行。
        因此不返回行 id，写入后从清单取。

        `price_multiplier` 缺省取 1；`request_overrides` 必须是 JSON 对象，
        转发时叠加到请求体上。请求方须持有管理面能力，越权返回 `code=403`。
      operationId: setAdminModelMap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetAdminModelMapRequest"
      responses:
        "200":
          description: 已写入。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/admin/accounts:
    get:
      tags: [admin]
      summary: 账户清单
      description: |
        列出全部账户，与 `tokenmp admin account list --json` 同一份行数据。
        含归属主体与默认商家的标识：管理面要据此把账户与运营对象对应起来。
        请求方须持有管理面能力，越权返回 `code=403`。
      operationId: listAdminAccounts
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: 账户清单。
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PageOfAdminAccount"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/admin/pricing:
    get:
      tags: [admin]
      summary: 定价版本清单
      description: |
        列出定价版本，与 `tokenmp admin price list --json` 同一份行数据；按 id 升序。
        未下架的行 `retired_at` 为 `null`，据此区分生效与已退役版本（与 CLI 的
        `active` / `retired` 列同一判定）。
        请求方须持有管理面能力，越权返回 `code=403`。
      operationId: listAdminPricing
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - $ref: "#/components/parameters/AdminMerchantFilter"
        - $ref: "#/components/parameters/ModelFilter"
      responses:
        "200":
          description: 定价版本清单。
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PageOfAdminPricing"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/admin/quotas:
    get:
      tags: [admin]
      summary: 窗口限额清单
      description: |
        列出窗口限额定义，附当前窗口的已用量与剩余额度，与
        `tokenmp admin quota list --json` 同一份行数据。已用量与剩余额度复用判定链上的
        同一份窗口计算；无法判定的行两者为 `null`，展示层给占位符。
        请求方须持有管理面能力，越权返回 `code=403`。
      operationId: listAdminQuotas
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - $ref: "#/components/parameters/AdminScopeFilter"
        - $ref: "#/components/parameters/AdminScopeIDFilter"
        - $ref: "#/components/parameters/AdminAccountFilter"
      responses:
        "200":
          description: 限额清单。
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PageOfAdminQuota"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/admin/adjustments:
    get:
      tags: [admin]
      summary: 调账清单
      description: |
        列出人工调账记录，与 `tokenmp admin adjust list --json` 同一份行数据；按 id 升序。
        调整数量是十进制字符串，正数为补扣、负数为退费。
        请求方须持有管理面能力，越权返回 `code=403`。
      operationId: listAdminAdjustments
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - $ref: "#/components/parameters/AdminAccountFilter"
      responses:
        "200":
          description: 调账清单。
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PageOfAdminAdjustment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/admin/usage:
    get:
      tags: [admin]
      summary: 全平台用量流水
      description: |
        列出用量流水，与 `tokenmp admin usage list --json` 同一份行数据；按 id 升序。
        与用户面流水的差别只在作用域：这里跨账户，用于对账与排障，因此多出
        `merchant_id` 与结算字段。
        请求方须持有管理面能力，越权返回 `code=403`。
      operationId: listAdminUsage
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - $ref: "#/components/parameters/AdminAccountFilter"
        - $ref: "#/components/parameters/Since"
      responses:
        "200":
          description: 用量流水。
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PageOfAdminUsageItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/admin/settlements:
    get:
      tags: [admin]
      summary: 全平台结算对账单
      description: |
        按商家逐行列出结算对账单，字段与 `tokenmp admin settlement list --json` 的行一致。
        `merchant_id` 限定单个商家；账期缺省时按各商家自己的账期取上一个完整自然周期 ——
        账期是商家自己的口径，同一批里各家的账期可以不同。
        请求方须持有管理面能力，越权返回 `code=403`。
      operationId: listAdminSettlements
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - name: merchant_id
          in: query
          required: false
          description: 按商家过滤；缺省列出全部商家。
          schema:
            type: integer
            format: int64
            minimum: 1
        - $ref: "#/components/parameters/WindowFrom"
        - $ref: "#/components/parameters/WindowTo"
      responses:
        "200":
          description: 对账单清单。
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PageOfAdminSettlementItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    webSession:
      type: http
      scheme: bearer
      description: |
        会话访问令牌，登录响应的 `data.access_token` 取得，随 `Authorization: Bearer` 提交。
        与数据面的模型 API 密钥是两套独立凭据，不得互换。
...
  schemas:
...
    Envelope:
      type: object
      description: 页面通信的统一信封，六个字段固定出现。
      required: [code, data, message, page, size, total]
      properties:
        code:
          $ref: "#/components/schemas/BusinessCode"
        data:
          description: 业务数据，结构由各端点声明；无数据时省略或为 null。
        message:
          type: string
          description: 面向用户的文案；程序不得按它分支。
        page:
          type: [integer, "null"]
          description: 当前页码，从 1 起；非列表端点为 null。
        size:
          type: [integer, "null"]
          description: 每页条数；非列表端点为 null。
        total:
          type: [integer, "null"]
          description: 总条数；非列表端点为 null。
...
    PageOfAdminChannel:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminChannel"
      additionalProperties: false
...
    PageOfAdminCredential:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminCredential"
      additionalProperties: false
...
    PageOfAdminModelMap:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminModelMap"
      additionalProperties: false
...
    PageOfAdminMerchant:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminMerchant"
      additionalProperties: false
    CreateAdminMerchantRequest:
      type: object
      description: 新建商家的请求体。
      required: [code, name, kind]
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 64
          description: 商家编码，平台内唯一；重复返 `code=409`。
        name:
          type: string
          minLength: 1
          maxLength: 128
        kind:
          type: string
          enum: [platform, partner]
      additionalProperties: false
...
    CreateAdminChannelRequest:
      type: object
      description: 新建渠道的请求体。
      required: [merchant_id, name, type, cred_group, base_url]
      properties:
        merchant_id:
          type: integer
          format: int64
          minimum: 1
          description: 归属商家 id，取自商家清单。
        name:
          type: string
          minLength: 1
          maxLength: 64
        vendor:
          type: string
          maxLength: 64
          description: 厂商标识，如 `openai` / `anthropic` / `gemini`；仅供展示与对账分组。
        type:
          type: string
          description: 渠道的线协议类型。
          enum: [openai_chat, openai_responses, anthropic_messages, gemini_generate]
        cred_group:
          type: string
          minLength: 1
          maxLength: 64
          description: 凭据分组名；本端点只建立分组名，凭据另行写入。
        base_url:
          type: string
          minLength: 1
          maxLength: 255
          description: 上游基地址，只填到端点段之前。
        credential_style:
          type: string
          description: 凭据注入形态；省略时按协议现状注入。
          enum: [authorization, x-api-key, x-goog-api-key, query]
        config:
          type: object
          additionalProperties: true
          description: |
            渠道级扩展配置原文：`headers` 是静态请求头（字符串到字符串的对象，不得占用
            网关自身的头名），其余键供探针声明。省略时无扩展配置。
        priority:
          type: integer
          minimum: 1
          description: 选路优先级，值小者优先；省略取 100。
        weight:
          type: integer
          minimum: 1
          description: 同优先级内的加权轮询权重；省略取 100。
      additionalProperties: false
    CreateAdminCredentialRequest:
      type: object
      description: 写入上游凭据的请求体。
      required: [merchant_id, cred_group, name, api_key]
      properties:
        merchant_id:
          type: integer
          format: int64
          minimum: 1
          description: 归属商家 id。
        cred_group:
          type: string
          minLength: 1
          maxLength: 64
          description: 凭据分组，与渠道的 `cred_group` 对应。
        name:
          type: string
          minLength: 1
          maxLength: 64
        api_key:
          type: string
          minLength: 1
          description: 上游凭据明文，只在写入这一条路径出现；响应不回显。
      additionalProperties: false
    SetAdminModelMapRequest:
      type: object
      description: 写入渠道模型映射的请求体；按 `(channel_id, model)` 幂等。
      required: [channel_id, model, upstream_model]
      properties:
        channel_id:
          type: integer
          format: int64
          minimum: 1
        model:
          type: string
          minLength: 1
          maxLength: 128
          description: 客户端请求的模型名。
        upstream_model:
          type: string
          minLength: 1
          maxLength: 128
          description: 转发给该渠道时替换成的模型名。
        price_multiplier:
          type: string
          description: 该映射的价格倍率，十进制字符串；省略取 1。
        request_overrides:
          type: object
          additionalProperties: true
          description: 转发时叠加到请求体上的字段；省略时不叠加。
      additionalProperties: false
    AdminResourceID:
      type: object
      description: 写入动作的结果：受影响行的主键。
      required: [id]
      properties:
        id:
          type: integer
          format: int64
      additionalProperties: false
...
    PageOfAdminAccount:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminAccount"
      additionalProperties: false
...
    PageOfAdminPricing:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminPricing"
      additionalProperties: false
...
    PageOfAdminQuota:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminQuota"
      additionalProperties: false
...
    PageOfAdminAdjustment:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminAdjustment"
      additionalProperties: false
...
  parameters:
    Page:
      name: page
      in: query
      required: false
      description: 页码，从 1 起；缺省 1。
      schema:
        type: integer
        minimum: 1
        default: 1
    Size:
      name: size
      in: query
      required: false
      description: 每页条数；缺省 20，上限 100。
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    ModelFilter:
      name: model
      in: query
      required: false
      description: 按模型名精确匹配；缺省不过滤。
      schema:
        type: string
...
    AdminMerchantFilter:
      name: merchant_id
      in: query
      required: false
      description: 按商家 id 过滤；缺省不过滤。
      schema:
        type: integer
        format: int64
        minimum: 1
    AdminAccountFilter:
      name: account_id
      in: query
      required: false
      description: 按账户 id 过滤；缺省不过滤。管理面独有：用户面的作用域一律由会话推导。
      schema:
        type: integer
        format: int64
        minimum: 1
    AdminScopeFilter:
      name: scope
      in: query
      required: false
      description: |
        按限额 / 规则的作用范围过滤，取值 `account` / `api_key` / `channel` / `plan`；
        与 `scope_id` 必须成对给出，缺省列出全部。
      schema:
        type: string
        enum: [account, api_key, channel, plan]
    AdminScopeIDFilter:
      name: scope_id
      in: query
      required: false
      description: 范围实体的 id，与 `scope` 成对使用。
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
    BadRequest:
      description: 参数缺失或非法，`code=400`。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    Unauthorized:
      description: 未登录、会话失效或凭据错误，`code=401`。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
...
    Forbidden:
      description: 已登录但无权访问该资源，`code=403`。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
...
    Conflict:
      description: 状态冲突，`code=409`。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    InternalError:
      description: 服务端错误，`code=500`。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"